Vario accoustic feedback
For each model, we store two values:
- vario sensor (sensor name, None to disable vario)
- vario gain   (0 .. 15)
These can be set in the Extra -> Vario submenu.

If there is no alarm at all, we check whether we have telemetry input,
and whether the sensor value has changed since two seconds ago.
When it did not change, we remain silent. When it did, we play a short
beep at base frequency (1200 Hz), and a longer beep representing
the difference in sensor value since the previous beep: increased value
means higher frequency, decreased value means lower frequency.

The gain value affects the sensor difference to frequency difference
mapping: bigger the gain, the bigger the frequency difference for a given
sensor value difference.

The frequency range is limited between 600 and 2400 Hz. When the difference
is negative, we use one step smaller gain, to account for half range
of frequencies in the negative side.

FIXME: when the sensor ID is uninitialized, the Tx sometimes reboots
when Rx is connected. Set the sensor to something sane or to None,
when this happens to you. User alarms have the same problem.