Vario setup: fix for max gain
When the gain value is at its maximum (15),
the UP keypress is incorrectly handed to the
next branch, thus increasing the Dead value.

A nested if() fixes this instead of &&.

Also reformatted the whole function to use
tabs instead of four spaces, and if () {}
with curly brackets on the same level
of indentation.